Move Drawer Menu Item from Stacks to navigation-options
Also clean up unused imports from Stack files
