+++ react-native/config/navigation-options.js
@@ -1,3 +1,7 @@
+import React from 'react';
+import { TouchableOpacity, StyleSheet } from 'react-native';
+import { Ionicons } from '@expo/vector-icons';
+
 import Colors from './colors';
 
 export const headerOptions = {
@@ -7,3 +11,25 @@ export const headerOptions = {
     },
 };
 
+export const drawerMenuIcon = navigation => {
+      return(
+        <TouchableOpacity
+          style={styles.drawerIcon}
+          onPress={() => {
+            navigation.toggleDrawer();
+          }}
+        >
+          <Ionicons
+            name='md-menu'
+            size={30}
+            color={Colors.light_shade4}
+          />
+        </TouchableOpacity>
+      );
+};
+
+const styles = StyleSheet.create({
+  drawerIcon: {
+    paddingLeft: 15,
+  },
+});
