+++ react-native/navigation/AboutStack.js
@@ -1,10 +1,10 @@
 import React, { useContext } from 'react';
-import { TouchableOpacity, StyleSheet } from 'react-native';
+import { StyleSheet } from 'react-native';
 import { createStackNavigator } from '@react-navigation/stack';
 import { Ionicons } from '@expo/vector-icons';
 
 import AboutScreen from '../screens/AboutScreen';
-import { headerOptions } from '../config/navigation-options';
+import { headerOptions, drawerMenuIcon } from '../config/navigation-options';
 import Colors from '../config/colors';
 
 const Stack = createStackNavigator();
@@ -15,22 +15,7 @@ const AboutStack = props => {
     const AboutScreenOptions = {
         title: 'About Common Goods',
         ...headerOptions,
-        headerLeft: () => {
-            return(
-                <TouchableOpacity
-                  style={styles.drawerIcon}
-                  onPress={() => {
-                      navigation.toggleDrawer();
-                  }}
-                >
-                  <Ionicons
-                    name='md-menu'
-                    size={30}
-                    color={Colors.light_shade4}
-                  />
-                </TouchableOpacity>
-            );
-        },
+      headerLeft: drawerMenuIcon.bind(this, navigation),
     };
 
     return(
@@ -47,9 +32,6 @@ const AboutStack = props => {
 };
 
 const styles = StyleSheet.create({
-    drawerIcon: {
-        paddingLeft: 15,
-    },
 });
 
 export default AboutStack;
